implemented core data function of note on favorites and pricese, also grab the fetch function out of origin findOrCreate function. Next: build note provider and Manager

+++ GoToMarket/GoToMarket/Model/CoreData/CropDatas+CoreDataProperties.swift
@@ -21,7 +21,7 @@ extension CropDatas {
     @NSManaged public var cropCode: String
     @NSManaged public var cropName: String
     @NSManaged public var marketName: String?
-    @NSManaged public var averagePrice: Float
+    @NSManaged public var averagePrice: Double
     @NSManaged public var note: UserNotes?
 
 }
